\name{test.dat.seq}
\alias{test.dat.seq}
\docType{data}
\title{
Testing Sequence Index Dataset
}
\description{
An example, a data frame of sequence index of testing samples
}
\usage{data(test.dat.seq)}
\format{
  A data frame with 1527 observations on the following 85 variables.
  \describe{
    \item{\code{gene}}{a factor with levels \code{ERCC-00002} \code{ERCC-00003} \code{ERCC-00007} \code{ERCC-00009} \code{ERCC-00012} \code{ERCC-00013} \code{ERCC-00014} \code{ERCC-00016} \code{ERCC-00017} \code{ERCC-00018} \code{ERCC-00019} \code{ERCC-00022} \code{ERCC-00023} \code{ERCC-00024} \code{ERCC-00025} \code{ERCC-00028} \code{ERCC-00031} \code{ERCC-00033} \code{ERCC-00034} \code{ERCC-00035} \code{ERCC-00039} \code{ERCC-00040} \code{ERCC-00041} \code{ERCC-00042} \code{ERCC-00044} \code{ERCC-00048} \code{ERCC-00054} \code{ERCC-00057} \code{ERCC-00058} \code{ERCC-00059} \code{ERCC-00061} \code{ERCC-00079} \code{ERCC-00084} \code{ERCC-00170}}
    \item{\code{pos}}{a numeric vector}
    \item{\code{code_strand}}{a factor with levels \code{+}}
    \item{\code{seq_strand}}{a factor with levels \code{S}}
    \item{\code{X1}}{a numeric vector}
    \item{\code{X2}}{a numeric vector}
    \item{\code{X3}}{a numeric vector}
    \item{\code{X4}}{a numeric vector}
    \item{\code{X5}}{a numeric vector}
    \item{\code{X6}}{a numeric vector}
    \item{\code{X7}}{a numeric vector}
    \item{\code{X8}}{a numeric vector}
    \item{\code{X9}}{a numeric vector}
    \item{\code{X10}}{a numeric vector}
    \item{\code{X11}}{a numeric vector}
    \item{\code{X12}}{a numeric vector}
    \item{\code{X13}}{a numeric vector}
    \item{\code{X14}}{a numeric vector}
    \item{\code{X15}}{a numeric vector}
    \item{\code{X16}}{a numeric vector}
    \item{\code{X17}}{a numeric vector}
    \item{\code{X18}}{a numeric vector}
    \item{\code{X19}}{a numeric vector}
    \item{\code{X20}}{a numeric vector}
    \item{\code{X21}}{a numeric vector}
    \item{\code{X22}}{a numeric vector}
    \item{\code{X23}}{a numeric vector}
    \item{\code{X24}}{a numeric vector}
    \item{\code{X25}}{a numeric vector}
    \item{\code{X26}}{a numeric vector}
    \item{\code{X27}}{a numeric vector}
    \item{\code{X28}}{a numeric vector}
    \item{\code{X29}}{a numeric vector}
    \item{\code{X30}}{a numeric vector}
    \item{\code{X31}}{a numeric vector}
    \item{\code{X32}}{a numeric vector}
    \item{\code{X33}}{a numeric vector}
    \item{\code{X34}}{a numeric vector}
    \item{\code{X35}}{a numeric vector}
    \item{\code{X36}}{a numeric vector}
    \item{\code{X37}}{a numeric vector}
    \item{\code{X38}}{a numeric vector}
    \item{\code{X39}}{a numeric vector}
    \item{\code{X40}}{a numeric vector}
    \item{\code{X41}}{a numeric vector}
    \item{\code{X42}}{a numeric vector}
    \item{\code{X43}}{a numeric vector}
    \item{\code{X44}}{a numeric vector}
    \item{\code{X45}}{a numeric vector}
    \item{\code{X46}}{a numeric vector}
    \item{\code{X47}}{a numeric vector}
    \item{\code{X48}}{a numeric vector}
    \item{\code{X49}}{a numeric vector}
    \item{\code{X50}}{a numeric vector}
    \item{\code{X51}}{a numeric vector}
    \item{\code{X52}}{a numeric vector}
    \item{\code{X53}}{a numeric vector}
    \item{\code{X54}}{a numeric vector}
    \item{\code{X55}}{a numeric vector}
    \item{\code{X56}}{a numeric vector}
    \item{\code{X57}}{a numeric vector}
    \item{\code{X58}}{a numeric vector}
    \item{\code{X59}}{a numeric vector}
    \item{\code{X60}}{a numeric vector}
    \item{\code{X61}}{a numeric vector}
    \item{\code{X62}}{a numeric vector}
    \item{\code{X63}}{a numeric vector}
    \item{\code{X64}}{a numeric vector}
    \item{\code{X65}}{a numeric vector}
    \item{\code{X66}}{a numeric vector}
    \item{\code{X67}}{a numeric vector}
    \item{\code{X68}}{a numeric vector}
    \item{\code{X69}}{a numeric vector}
    \item{\code{X70}}{a numeric vector}
    \item{\code{X71}}{a numeric vector}
    \item{\code{X72}}{a numeric vector}
    \item{\code{X73}}{a numeric vector}
    \item{\code{X74}}{a numeric vector}
    \item{\code{X75}}{a numeric vector}
    \item{\code{X76}}{a numeric vector}
    \item{\code{X77}}{a numeric vector}
    \item{\code{X78}}{a numeric vector}
    \item{\code{X79}}{a numeric vector}
    \item{\code{X80}}{a numeric vector}
    \item{\code{X81}}{a numeric vector}
  }
}
\references{
Cai G, RNA-SEQUENCING APPLICATIONS: GENE EXPRESSION QUANTIFICATION AND METHYLATOR PHENOTYPE IDENTIFICATION, Ph.D. Thesis, 2013
}
\examples{
data(test.dat.seq)
}
\keyword{datasets}
