\name{train.dat.counts}
\alias{train.dat.counts}
\docType{data}
\title{
Training Counts Dataset
}
\description{
An example, a data frame of reads counts of training samples
}
\usage{data(train.dat.counts)}
\format{
  A data frame with 5203 observations on the following 6 variables.
  \describe{
    \item{\code{gene}}{a factor with levels \code{ERCC-00002} \code{ERCC-00003} \code{ERCC-00007} \code{ERCC-00009} \code{ERCC-00012} \code{ERCC-00013} \code{ERCC-00014} \code{ERCC-00016} \code{ERCC-00017} \code{ERCC-00018} \code{ERCC-00019} \code{ERCC-00022} \code{ERCC-00023} \code{ERCC-00024} \code{ERCC-00025} \code{ERCC-00028} \code{ERCC-00031} \code{ERCC-00033} \code{ERCC-00034} \code{ERCC-00035} \code{ERCC-00039} \code{ERCC-00040} \code{ERCC-00041} \code{ERCC-00042} \code{ERCC-00044} \code{ERCC-00048} \code{ERCC-00054} \code{ERCC-00058} \code{ERCC-00061} \code{ERCC-00079} \code{ERCC-00171}}
    \item{\code{pos}}{a numeric vector}
    \item{\code{code_strand}}{a factor with levels \code{+}}
    \item{\code{seq_strand}}{a factor with levels \code{S}}
    \item{\code{GSM516589_run32_s_2_ERCC_map_Pos}}{a numeric vector}
    \item{\code{GSM517059_run29_s_1_ERCC_map_Pos}}{a numeric vector}
  }
}
\references{
Cai G, RNA-SEQUENCING APPLICATIONS: GENE EXPRESSION QUANTIFICATION AND METHYLATOR PHENOTYPE IDENTIFICATION, Ph.D. Thesis, 2013
}
\examples{
data(train.dat.counts)
}
\keyword{datasets}
